feat!: widen Usage token counters to u64
Usage is also an accumulator (run totals, UsageSink across sub-agents,
application-side billing), and u32::MAX is reachable in those totals.
Adapter wire types and OnlineCalibratedTokenizer::observe move to u64
so the mapping needs no casts.

+++ adapters/anthropic/src/stream.rs
@@ -191,7 +191,7 @@ where
 /// is strictly more informative.
 fn apply_cache_creation(
     usage: &mut Usage,
-    flat_input_tokens: u32,
+    flat_input_tokens: u64,
     breakdown: Option<&crate::events::CacheCreationBreakdown>,
 ) {
     match breakdown {
