+++ ailoop/src/engine.rs
@@ -1838,7 +1838,7 @@ mod tests {
         }
     }
 
-    fn tokens(input: u32, output: u32) -> Usage {
+    fn tokens(input: u64, output: u64) -> Usage {
         let mut u = Usage::default();
         u.input_tokens = input;
         u.output_tokens = output;
